+++ frontend/src/lib/components/ExportButton/ExportButton.tsx
@@ -1,50 +1,30 @@
 import React from 'react'
+import { ExporterFormat } from '~/types'
 import { LemonButton, LemonButtonProps, LemonButtonWithPopup } from '../LemonButton'
-import { useActions, useValues } from 'kea'
-import { ExporterFormat, exporterLogic } from './exporterLogic'
 import { LemonDivider } from '../LemonDivider'
-import { insightLogic } from 'scenes/insights/insightLogic'
-import { InsightLogicProps, InsightShortId } from '~/types'
+import { triggerExport } from './exporter'
 
-interface ExportButtonProps extends Pick<LemonButtonProps, 'icon' | 'type' | 'fullWidth'> {
-    dashboardId?: number
-    insightShortId?: InsightShortId
+export interface ExportButtonItemResource {
+    method?: 'GET' | 'POST'
+    path: string
+    body?: any
+    filename?: string
 }
 
-export function ExportButton({ dashboardId, insightShortId, ...buttonProps }: ExportButtonProps): JSX.Element {
-    const insightLogicProps: InsightLogicProps = {
-        dashboardItemId: insightShortId,
-        doNotLoad: true,
-    }
+export interface ExportButtonItem {
+    title?: string
+    format: ExporterFormat
+    resource: ExportButtonItemResource
+}
 
-    const { supportsCsvExport, csvExportUrl, insight } = useValues(insightLogic(insightLogicProps))
+export interface ExportButtonProps extends Pick<LemonButtonProps, 'icon' | 'type' | 'fullWidth'> {
+    items: ExportButtonItem[]
+}
 
-    const { exportItem } = useActions(exporterLogic({ dashboardId, insightId: insight?.id }))
-    const { exportInProgress } = useValues(exporterLogic({ dashboardId, insightId: insight?.id }))
-
-    const supportedFormats: ExporterFormat[] = []
-
-    if (dashboardId || insightShortId) {
-        supportedFormats.push(ExporterFormat.PNG)
-    }
-    if (supportsCsvExport) {
-        supportedFormats.push(ExporterFormat.CSV)
-    }
-
-    const onExportItemClick = (exportFormat: ExporterFormat): void => {
-        // NOTE: Once we standardise the exporting code in the backend this can be removed
-        if (exportFormat === ExporterFormat.CSV) {
-            window.open(csvExportUrl, '_blank')
-            return
-        }
-
-        exportItem(exportFormat)
-    }
-
+export function ExportButton({ items, ...buttonProps }: ExportButtonProps): JSX.Element {
     return (
         <LemonButtonWithPopup
             type="stealth"
-            loading={exportInProgress}
             data-attr="export-button"
             {...buttonProps}
             popup={{
@@ -54,15 +34,20 @@ export function ExportButton({ dashboardId, insightShortId, ...buttonProps }: Ex
                     <>
                         <h5>File type</h5>
                         <LemonDivider />
-                        {supportedFormats.map((format) => (
+                        {items.map((item) => (
                             <LemonButton
-                                key={format}
+                                key={`${item.format}-${item.resource.path}`}
                                 fullWidth
                                 type="stealth"
-                                onClick={() => onExportItemClick(format)}
-                                data-attr={`export-button-${format.split('/').pop()}`}
+                                onClick={() =>
+                                    triggerExport({
+                                        export_format: item.format,
+                                        export_context: item.resource,
+                                    })
+                                }
+                                data-attr={`export-button-${item.format.split('/').pop()}`}
                             >
-                                .{format.split('/').pop()}
+                                {item.title ? item.title : `.${item.format.split('/').pop()}`}
                             </LemonButton>
                         ))}
                     </>
