+++ frontend/src/scenes/insights/Insight.tsx
@@ -4,7 +4,7 @@ import { useActions, useMountedLogic, useValues, BindLogic } from 'kea'
 import { insightSceneLogic } from 'scenes/insights/insightSceneLogic'
 import { insightLogic } from './insightLogic'
 import { insightCommandLogic } from './insightCommandLogic'
-import { ItemMode, AvailableFeature, InsightShortId, InsightModel, InsightType } from '~/types'
+import { ItemMode, AvailableFeature, InsightShortId, InsightModel, InsightType, ExporterFormat } from '~/types'
 import { NPSPrompt } from 'lib/experimental/NPSPrompt'
 import { SaveCohortModal } from 'scenes/trends/SaveCohortModal'
 import { personsModalLogic } from 'scenes/trends/personsModalLogic'
@@ -27,7 +27,6 @@ import { LemonButton } from 'lib/components/LemonButton'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { EditorFilters } from './EditorFilters/EditorFilters'
-import { ExportButton } from 'lib/components/ExportButton/ExportButton'
 import { More } from 'lib/components/LemonButton/More'
 import { LemonDivider } from 'lib/components/LemonDivider'
 import { deleteWithUndo } from 'lib/utils'
@@ -39,8 +38,7 @@ import { SubscriptionsModal, SubscribeButton } from 'lib/components/Subscription
 import { UserActivityIndicator } from 'lib/components/UserActivityIndicator/UserActivityIndicator'
 import clsx from 'clsx'
 import { SharingModal } from 'lib/components/Sharing/SharingModal'
-import { ExportButtonV2 } from 'lib/components/ExportButton/ExportButtonV2'
-import { ExporterFormat } from 'lib/components/ExportButton/exporter'
+import { ExportButton } from 'lib/components/ExportButton/ExportButton'
 
 export function Insight({ insightId }: { insightId: InsightShortId | 'new' }): JSX.Element {
     const { insightMode, subscriptionId } = useValues(insightSceneLogic)
@@ -175,9 +173,8 @@ export function Insight({ insightId }: { insightId: InsightShortId | 'new' }): J
                                                     {usingSubscriptionFeature && (
                                                         <SubscribeButton insightShortId={insight.short_id} />
                                                     )}
-                                                    <ExportButton insightShortId={insight.short_id} fullWidth />
                                                     {exporterResourceParams ? (
-                                                        <ExportButtonV2
+                                                        <ExportButton
                                                             fullWidth
                                                             items={[
                                                                 {
